첨부파일 기능 추가.

+++ src/main/java/kr/or/ddit/lecture/assignment/service/AssignmentSubServiceImpl.java
@@ -1,12 +1,22 @@
 package kr.or.ddit.lecture.assignment.service;
 
+import java.io.IOException;
 import java.text.MessageFormat;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.UUID;
 
 import javax.inject.Inject;
 
+import org.apache.commons.lang.StringUtils;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.core.io.Resource;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
 
+import kr.or.ddit.atch.service.AtchFileService;
+import kr.or.ddit.atch.vo.AtchFileDetailVO;
+import kr.or.ddit.atch.vo.AtchFileVO;
 import kr.or.ddit.exception.PKNotFoundException;
 import kr.or.ddit.lecture.assignment.dao.AssignmentSubDao;
 import kr.or.ddit.lecture.assignment.vo.AssignmentSubVO;
@@ -16,6 +26,12 @@ public class AssignmentSubServiceImpl implements AssignmentSubService {
 
 	@Inject
 	private AssignmentSubDao assignmentSubDao;
+	
+	@Inject
+	private AtchFileService atchService; 
+	
+	@Value("#{appInfo.asubPath}")
+	private Resource asubPath;
 	
 	@Override
 	public List<AssignmentSubVO> findListByAssignment(String asNo) {
@@ -34,19 +50,63 @@ public class AssignmentSubServiceImpl implements AssignmentSubService {
 		return assignmentSub; 
 	}
 	
+	private void processAtchFileGroup(AssignmentSubVO asub) {
+		MultipartFile[] subFiles = asub.getSubFiles();
+		if (subFiles == null) return;
+		List<AtchFileDetailVO> detailList = new ArrayList<>();
+		for (MultipartFile subFile : subFiles) {
+			if (subFile.isEmpty()) continue;
+			detailList.add(new AtchFileDetailVO(subFile));
+		}
+		if (detailList.size() > 0) {
+			AtchFileVO fileGroup = new AtchFileVO();
+			fileGroup.setDetailList(detailList);
+			// 1. 첨부파일의 메타 데이터 저장
+			// 2. 첨부파일의 2진 데이터 저장
+			try {
+				if (StringUtils.isNotBlank(asub.getAsubFile())) {
+					// 운항일지 수정인경우 첨부파일 데이터 전부 삭제하고 다시 생성
+					atchService.removeAtchFileGroup(asub.getAsubFile(), asubPath);
+				}
+				// 운항일지 작성
+				fileGroup.setAtchFileId(UUID.randomUUID().toString().substring(0, 20));
+				atchService.createAtchFileGroup(fileGroup, asubPath);
+				asub.setAsubFile(fileGroup.getAtchFileId());
+			} catch (IOException e) {
+				throw new RuntimeException(e);
+			}
+		}
+	}
+	
 	@Override
-	public boolean create(AssignmentSubVO obj) {
-		return assignmentSubDao.insert(obj) > 0;
+	public boolean create(AssignmentSubVO asub) {
+		processAtchFileGroup(asub);
+		return assignmentSubDao.insert(asub) > 0;
 	}
 
 	@Override
-	public boolean modify(AssignmentSubVO obj) {
-		return assignmentSubDao.update(obj) > 0;
+	public boolean modify(AssignmentSubVO asub) {
+		processAtchFileGroup(asub);
+		return assignmentSubDao.update(asub) > 0;
 	}
 
 	@Override
 	public boolean removeByAsNoAndStdNo(AssignmentSubVO asub) {
-		return assignmentSubDao.deleteByAsNoAndStdNo(asub) > 0;
+		AssignmentSubVO findEntity = assignmentSubDao.selectByIds(asub);
+		// 만약 파일이 존재하면 이것부터 지우기
+		boolean success = true;
+		if (findEntity.getAsubFile() != null) {
+			try {
+				success = atchService.removeAtchFileGroup(findEntity.getAsubFile(), asubPath);
+			} catch (IOException e) {
+				throw new RuntimeException(e);
+			}
+		}
+		// 성공적으로 파일지웠으면 이제 제출데이터 지우기
+		if (success) {
+			return assignmentSubDao.deleteByAsNoAndStdNo(asub) > 0;
+		}
+		return false;
 	}
 	
 	@Override
